JuliaHashcode

Documentation for JuliaHashcode.

JuliaHashcode.Car

JuliaHashcode.Solution

JuliaHashcode.find_best_path

JuliaHashcode.get_street_end

JuliaHashcode.greedy_algo

JuliaHashcode.is_feasible

JuliaHashcode.is_street

JuliaHashcode.is_street_start

JuliaHashcode.read_problem

JuliaHashcode.read_solutions

JuliaHashcode.total_distance

JuliaHashcode.update_visited_path

JuliaHashcode.write_city

JuliaHashcode.write_solutions

JuliaHashcode.Car — Type

Car

Store a car made of the car's path, duration, and if it's still running.

Fields

path::Vector{Int}: path the car has traveled

duration::Int: total duration the car has traveled

is_running::Bool: tells if the car is still running or reached end of duration

JuliaHashcode.Solution — Type

Solution

Store a set of itineraries, one for each car.

Fields

city::City:

itineraries::Vector{Vector{Int}}: each itinerary is a vector of junction indices

feasible::Bool: tells if the solution is feasible given the city

total::Int: total is the total distance the cars travel in the solution

JuliaHashcode.find_best_path — Method

find_best_path(depth, total_depth, duration, total_duration, streets, visited, neighboring_streets, current_position)

Runs a recursive greedy algorithm to find the best path of length total_depth from current_position making sure it doesn't exceed duration

JuliaHashcode.get_street_end — Method

get_street_end(i, street)

Retrieve the arrival endpoint of street when it starts at junction i.

JuliaHashcode.greedy_algo — Method

greedy_algo(problem)

Runs a greedy algorithm for each car to find the best paths to achieve the highest score.

JuliaHashcode.is_feasible — Method

is_feasible(solution, city[; verbose=false]) Check if solution satisfies the constraints for the instance defined by city. The following criteria are considered (taken from the problem statement):

the number of itineraries has to match the number of cars of city

the first junction of each itinerary has to be the starting junction of city

for each consecutive pair of junctions on an itinerary, a street connecting these junctions has to exist in city (if the street is one directional, it has to be traversed in the correct direction)

the duration of each itinerary has to be lower or equal to the total duration of city

JuliaHashcode.is_street — Method

is_street(i, j, street)

Check if the trip from junction i to junction j corresponds to a valid direction of street.

JuliaHashcode.is_street_start — Method

is_street_start(i, street)

Check if junction i corresponds to a valid starting point of street.

JuliaHashcode.read_problem — Function

read_problem(path)

Read and parse a City from a file located at path.

The default path is an artifact containing the official challenge data.

JuliaHashcode.read_solutions — Method

read_solution(solution, path) Read and parse a Solution from a file located at path.

JuliaHashcode.total_distance — Method

total_distance(solution, city)

Compute the total distance of all itineraries in solution based on the street data from city. Streets visited several times are only counted once.

JuliaHashcode.update_visited_path — Method

update_visited_path(best_path, visited, streets)

After finding the best_path, update visited to reflect the streets that has been traversed.

JuliaHashcode.write_city — Method

write_city(city, path)

Write a City to a file located at path.

JuliaHashcode.write_solutions — Method

write_solution(solution, path) Write a Solution to a file located at path.